chore: ディレクトリ構成を変更
