--- app/.server/loaders/index.ts
@@ -1,3 +1,0 @@
-import { getAvatarRanking } from './getAvatarRanking'
-import { getClothRanking } from './getClothRanking'
-export { getAvatarRanking, getClothRanking }
