+++ app/components/card/index.ts
@@ -1,4 +1,5 @@
 import { AvatarCard } from './ranking-avatar-card'
 import { ClothCard } from './ranking-cloth-card'
 import { TopRankingCard } from './top-ranking-card'
-export { AvatarCard, ClothCard, TopRankingCard }
+import { TopTrendCard } from './top-trend-card'
+export { AvatarCard, ClothCard, TopRankingCard ,TopTrendCard}
