+++ app/components/card/ranking-avatar-card.tsx
@@ -1,7 +1,7 @@
 import { Avatar, AvatarFallback, AvatarImage } from '~/components/ui/avatar'
 import { Card, CardContent, CardFooter, CardTitle } from '~/components/ui/card'
-import { excludeOldData } from '~/lib/date'
-import { getImageUrl, getShopImageUrl, truncateString } from '~/lib/utils'
+import { buildAvatarImage, buildShopImage, excludeOldDate } from '~/lib/format'
+import { truncateString } from '~/lib/utils'
 import type { RankingAvatarType } from '~/types/items'
 import { FavoriteTag } from '../element/favorite-tag'
 import { RankingTag } from '../element/ranking-tag'
@@ -24,7 +24,7 @@ export const AvatarCard = ({
 					</div>
 					<img
 						className="rounded-md"
-						src={getImageUrl(item.avatar_image)}
+						src={buildAvatarImage(item.avatar_image)}
 						loading="lazy"
 						alt={item.avatar_name}
 					/>
@@ -42,7 +42,7 @@ export const AvatarCard = ({
 				<div className="flex items-center gap-2">
 					<Avatar>
 						<AvatarImage
-							src={getShopImageUrl(item.shop_image)}
+							src={buildShopImage(item.shop_image)}
 							loading="lazy"
 							alt={item.shop_name}
 						/>
@@ -50,7 +50,7 @@ export const AvatarCard = ({
 					</Avatar>
 					<div className="pl-1 text-sm">{item.shop_name}</div>
 				</div>
-				<div>{excludeOldData(item.avatar_added)}</div>
+				<div>{excludeOldDate(item.avatar_added)}</div>
 			</CardFooter>
 		</Card>
 	)
