+++ app/components/element/top-cloth.tsx
@@ -1,6 +1,7 @@
 import { Avatar, AvatarFallback, AvatarImage } from '~/components/ui/avatar'
 import { Card, CardContent, CardFooter, CardTitle } from '~/components/ui/card'
-import { getImageUrl, getShopImageUrl, truncateString } from '~/lib/utils'
+import { buildAvatarImage, buildShopImage } from '~/lib/format'
+import { truncateString } from '~/lib/utils'
 import type { RankingClothType } from '~/types/items'
 import { FavoriteTag } from './favorite-tag'
 import { RankingTag } from './ranking-tag'
@@ -22,7 +23,7 @@ export const TopCloth = ({
 					</div>
 					<img
 						className="rounded-md"
-						src={getImageUrl(item.cloth_image)}
+						src={buildAvatarImage(item.cloth_image)}
 						loading="lazy"
 						alt={item.cloth_name}
 					/>
@@ -38,7 +39,7 @@ export const TopCloth = ({
 				<div className="flex items-center gap-2">
 					<Avatar>
 						<AvatarImage
-							src={getShopImageUrl(item.shop_image)}
+							src={buildShopImage(item.shop_image)}
 							loading="lazy"
 							alt={item.shop_name}
 						/>
