--- app/consts/urls.ts
@@ -1,11 +1,0 @@
-const ItemUrl = {
-	IMAGE_URL: 'https://booth.pximg.net/',
-	IMAGE_SIZE: '_base_resized.jpg',
-}
-
-const ShopUrl = {
-	IMAGE_URL: 'https://booth.pximg.net/c/128x128/users/',
-	IMAGE_SIZE: '_base_resized.jpg',
-}
-
-export { ItemUrl, ShopUrl }
