+++ app/lib/constants/urls.ts
@@ -1,0 +1,5 @@
+export const URLS = {
+	SHOP_IMAGE: 'https://booth.pximg.net/c/128x128/users/',
+	ITEM_IMAGE: 'https://booth.pximg.net/',
+	BASE_SIZE: '_base_resized.jpg',
+}
