+++ app/module/get/get-cloth-ranking.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { getAvatarRanking } from './index'
+import { getClothRanking } from './get-cloth-ranking'
 
 // cloudflare環境変数を含むcontextオブジェクトを作成
 type Context = {
@@ -16,20 +16,20 @@ const context: Context = {
 	},
 }
 
-describe('getAvatarRanking関数のテスト', () => {
+describe('getClothRanking関数のテスト', () => {
 	it('ランキングが取得できるかテスト', async () => {
 		const type = 'day'
 		const page = 1
 		const date = '2024-08-20'
 
-		const result = await getAvatarRanking(type, page, context, date)
+		const result = await getClothRanking(type, page, context, date)
 
 		expect(result.data).toBeDefined()
 		if (result.data) {
 			expect(Array.isArray(result.data)).toBe(true)
 			expect(result.data.length).toBeGreaterThan(0)
 			expect(result.data[0]).toHaveProperty('booth_id')
-			expect(result.data[0]).toHaveProperty('avatar_name')
+			expect(result.data[0]).toHaveProperty('cloth_name')
 		} else {
 			fail('result.data is null')
 		}
@@ -40,7 +40,7 @@ describe('getAvatarRanking関数のテスト', () => {
 		const page = 1
 		const date = '2023-08-26'
 
-		const result = await getAvatarRanking(type, page, context, date)
+		const result = await getClothRanking(type, page, context, date)
 
 		expect(result.data).toBeDefined()
 		if (result.data) {
@@ -57,7 +57,7 @@ describe('getAvatarRanking関数のテスト', () => {
 		const date = '2023-08-20'
 
 		try {
-			await getAvatarRanking(type, page, context, date)
+			await getClothRanking(type, page, context, date)
 		} catch (error) {
 			expect(error).toBeInstanceOf(Error)
 			expect((error as Error).message).toContain('Invalid ranking type')
