+++ app/module/get/get-ranking.ts
@@ -1,0 +1,4 @@
+import { getAvatarRanking } from './get-avatar-ranking'
+import { getClothRanking } from './get-cloth-ranking'
+
+export { getAvatarRanking, getClothRanking }
